Move back to Panda CSS + ConnectKit integration
This commit moves back from StyleX to Panda CSS.

StyleX is exclusively relying on a babel transform, which causes issues
with new tools and libraries where the general consensus is to avoid
relying on Babel, mostly for performances reasons.

Next has a Babel mode which I activated for StyleX, but it comes with a
few downsides:

- The compilation becomes much slower than with swc [1].
- I encountered one feature missing in this mode [2].
- It can create some tricky issues with external libraries, like
  RainbowKit and ConnectKit (which both embed client-side libraries
  relying on Node APIs).

The StyleX team is currently working on a solution that would allow to
not rely on Babel anymore [3], but it is not ready yet. For these
reasons and rather than spending too much time on these issues, I think
Panda would be the better option for SSR-compatible styles.

[1] https://swc.rs/
[2] https://nextjs.org/docs/messages/babel-font-loader-conflict
[3] https://github.com/facebook/stylex/pull/412