fix: don't panic when attempting to withdraw/repay too much
There was a `require()` that was intended to check for too much repayment,
but it was never being hit, as we were already panicking from arithmetic
overflow inside `_getNewICRFromTroveChange()`.

There was a late assertion to make sure we were never withdrawing more
than the Trove's entire collateral. Instead, we can do the check earlier
and make it a `require()` for a more helpful error message.