feat: include upfront fee incurred by batch in `BatchUpdated`
To reconstruct a batched Trove's transaction history, we have
to take into account not only the actions of the owner (and
add/remove-managers) but also the actions of the batch manager.

If the batch manager makes a premature interest rate adjustment,
the entire batch incurs an upfront fee. We include this fee in
the `BatchUpdated` event so that the individual fees incurred
by each Trove within the batch can be calculated.