fix: zombie Trove pointer reset despite no new zombie
Normally, `lastZombieTroveId` is first in line to get redeemed, however,
it will be skipped if its ICR is less than 100%. Previously, we would have
reset `lastZombieTroveId` to 0 if a subsequent redemption _doesn't_ result
in a Trove having `0 < debt < MIN_DEBT`. However, there's no reason to
reset `lastZombieTroveId` in this case, even if the chances of it escaping
liquidation until its ICR recovers to >100% are slim.