App: update TransactionFlow to add a postCheck() status
This allows to add an additional check that will ensure the change has
been propagated (generally to the indexer) before showing the success
screen. We also invalidate the queryClient cache at this stage, ensuring
that the data will get reload on the next screen.

This commit also adds a transition on the success message.