Loan Screen: fully redeemed state (#643)
A fully redeemed loan is determined from the combination of a redeemed
status (see subgraph to know how this is determined) and a debt of 0.

In this mode, the loan card only shows three values: the collateral (or
net value in leverage mode), the interest rate, and the redemption risk.
The three others will only get displayed if the loan debts goes above 0
again.