Merge the earnDeposit & earnWithdraw tx flows into a single earnUpdate
The new tx flow also makes sure the rewards are displayed in both cases.
If a reward is zero, it is not displayed.

The tx flow also make sure to handle the case of a new deposit (showing
the new values on the summary card) vs. an existing one (showing the
diff between the prev and new positions).