Earn position: fix the active / inactive state
The earn positions were considered active when the deposit was zero.
But it is possible to have non-zero rewards, while having a zero
deposit, e.g. in case of liquidations.

isEarnPositionActive() is now used, which ensures that an Earn position
is active if one of the three values (deposit / rewards.bold /
rewards.coll) is not zero.