Fetch troves from the the contracts rather than the subgraph
This changes makes fetching loans data more reliable by always using
getLatestTroveData(), and less dependent on the subgraph.

The subgraph is still used to fetch "trove statuses", which are
queries containing information that can’t be fetched by querying the
contracts:

- id
- closedAt
- createdAt
- mightBeLeveraged
- status

In the case of a single trove, the subgraph query is allowed to fail, in
which case defaults are provided.

Also:

- useLoansByAccount() and useLoanById() have moved to liquity-utils.ts.
- PositionLoanCommitted["updatedAt"] was unused and has been removed.