Subgraph: add previousOwner (#968)
When the NFT gets transferred to 0x0, knowing if this is part of a
liquidation or a normal transfer is difficult, since the transfer event
happens before the TroveOperation event.

So instead of trying to detect that a transfer is part of a liquidation,
previousOwner always reflects the previous owner before the last
transfer.

This will allow to query the liquidated troves of given account by
querying `{ status: liquidated, previousOwner: $account }`.